@@ -1,15 +1,15 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <RootNamespace>YL.NetCoreApp</RootNamespace>
     <AssemblyName>YL.NetCoreApp</AssemblyName>
     <LangVersion>7.1</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Diagnostics" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Server.Kestrel.Core" Version="2.2.0" />
+    <None Remove="DI\PropertiesAutowiredControllerActivator.txt" />
+    <None Remove="DI\PropertiesAutowiredFilterProvider.txt" />
   </ItemGroup>
 
   <ItemGroup>
