@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.Internal;
+
+//using Microsoft.AspNetCore.Http.Internal;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.AspNetCore.WebUtilities;
 using Newtonsoft.Json;
@@ -85,7 +86,8 @@ namespace YL.Utils.Http
         public static (string, string, string) ReadResultExecutingContext(this ResultExecutingContext context)
         {
             var req = context.HttpContext.Request;
-            req.EnableRewind();
+            //req.EnableRewind();
+            req.EnableBuffering();
             var method = req.Method;
             var url = $"{req.Scheme}://{req.Host}{req.PathBase}{req.Path}{req.QueryString}";
             var urlParam = req.QueryString.ToUriComponent();
@@ -116,7 +118,8 @@ namespace YL.Utils.Http
         public static string ReadFromResultExecutingContext(this ResultExecutingContext context)
         {
             var req = context.HttpContext.Request;
-            req.EnableRewind();
+            //req.EnableRewind();
+            req.EnableBuffering();
             var method = req.Method;
             var url = $"{req.Scheme}://{req.Host}{req.PathBase}{req.Path}{req.QueryString}";
             var urlParam = req.QueryString.ToUriComponent();
@@ -157,7 +160,8 @@ namespace YL.Utils.Http
         public static string ReadFromResultExecutedContext(this ResultExecutedContext context)
         {
             var req = context.HttpContext.Request;
-            req.EnableRewind();
+            //req.EnableRewind();
+            req.EnableBuffering();
             using (var ms = new MemoryStream())
             {
                 req.Body.Position = 0;
